Split error logging events into separate files
